Moving auth config task into its own class
